+++ beetroot/comp.py
@@ -5,10 +5,15 @@ import sys
 from .exception import *
 from .objtype import objtype
 
-class compr:
-    def strcompress(self, str_):
+class comp:
+    def compress(self, str_):
         """Compresses a strings with hybrid zlib/lzma"""
-        a = str(str_)
+        if objtype(str_) == "bytes":
+            a = str_.decode("iso-8859-1")
+            
+        else:
+            a = str(str_)
+            
         aa = "n" + a
         ab = "z" + zlib.compress(a.encode("utf-8")).decode("iso-8859-1")
         ac = "l" + lzma.compress(a.encode("utf-8")).decode("iso-8859-1")
@@ -18,8 +23,14 @@ class compr:
         
         return yay[yay2.index(min(yay2))]
     
-    def strdecompress(self, str_):
+    def decompress(self, str_):
         """Decompresses a strings with hybrid zlib/lzma"""
+        if objtype(str_) == "bytes":
+            str_ = str_.decode("iso-8859-1")
+            
+        else:
+            str_ = str(str_)
+            
         if str(str_).startswith("n"):
             return str(str_)[1:]
         
@@ -32,38 +43,4 @@ class compr:
         else:
             raise StringError("This string could not be properly decompressed.")
         
-    def bytecompress(self, b):
-        """Compresses a strings with hybrid zlib/lzma"""
-        if objtype(b) != "bytes":
-            raise InvalidTypeError("argument \"b\" must be bytestring")
-        
-        a = b
-        
-        aa = str("n" + a.decode("iso-8859-1")).encode("iso-8859-1")
-        ab = str("z" + zlib.compress(a).decode("iso-8859-1")).encode("iso-8859-1")
-        ac = str("l" + lzma.compress(a).decode("iso-8859-1")).encode("iso-8859-1")
-        
-        yay = [aa, ab, ac]
-        yay2 = [sys.getsizeof(aa), sys.getsizeof(ab), sys.getsizeof(ac)]
-        
-        return yay[yay2.index(min(yay2))]
-    
-    def bytedecompress(self, b):
-        """Decompresses a strings with hybrid zlib/lzma"""
-        if objtype(b) != "bytes":
-            raise InvalidTypeError("argument \"b\" must be bytestring")
-        
-        if b.decode("iso-8859-1").startswith("n"):
-            return str(b.decode("iso-8859-1"))[1:].encode("iso-8859-1")
-        
-        elif b.decode("iso-8859-1").startswith("z"):
-            return zlib.decompress(b.decode("iso-8859-1")[1:].encode("iso-8859-1"))
-        
-        elif b.decode("iso-8859-1").startswith("l"):
-            return lzma.decompress(b.decode("iso-8859-1")[1:].encode("iso-8859-1"))
-        
-        else:
-            raise StringError("This string could not be properly decompressed.")
-        
-comp = compr()
-del compr
+comp = comp()